Revert "fix: 170 - DEREFERENCE_NULLABLE - final int idx = unitsLRU[i]; - dereferenced expression unitsLRU is @Nullable"
This reverts commit fbbf476f12a17e08e7d05377803f6d4ca2e44648.

+++ gdx/src/com/badlogic/gdx/graphics/g3d/utils/DefaultTextureBinder.java
@@ -162,7 +162,7 @@ public final class DefaultTextureBinder implements TextureBinder {
   private final int bindTextureLRU(final GLTexture texture) {
     int i;
     for (i = 0; i < count; i++) {
-      final int idx = unitsLRU[i].intValue();
+      final int idx = unitsLRU[i];
       if (textures[idx] == texture) {
         reused = true;
         break;
@@ -172,12 +172,12 @@ public final class DefaultTextureBinder implements TextureBinder {
       }
     }
     if (i >= count) i = count - 1;
-    final int idx = unitsLRU[i].intValue();
+    final int idx = unitsLRU[i];
     while (i > 0) {
       unitsLRU[i] = unitsLRU[i - 1];
       i--;
     }
-    unitsLRU[0] = Integer.valueOf(idx);
+    unitsLRU[0] = idx;
     if (!reused) {
       textures[idx] = texture;
       texture.bind(offset + idx);
